Update LICENSE from AGPLv3 to Apache 2.0 (#17)
* Update LICENSE from AGPLv3 to Apache 2.0

After reconsiderations and getting some feedback our initial choice of license was too copyleft, we’ve decided to change it to Apache 2.0 before accepting any external contributions which will make changing license more difficult. 

As of now, all commits are from Boost employees , so we can have explicit approval of all commuters on that PR and from the CEO too. 

Signed-off-by: François Proulx <76956526+fproulx-boostsecurity@users.noreply.github.com>

* Update README.md

Signed-off-by: François Proulx <76956526+fproulx-boostsecurity@users.noreply.github.com>

* Update default.html

Signed-off-by: François Proulx <76956526+fproulx-boostsecurity@users.noreply.github.com>

---------

Signed-off-by: François Proulx <76956526+fproulx-boostsecurity@users.noreply.github.com>